Emacs: Initial config
Let's give it a try, finally! Even if I just pull some of these
experiences over into Neovim.

So far, I've found the `compile` command quite powerful, especially with
its compatibility with tools like grep. This is very similar to Vim's
makeprg and :make (with quickfix lists), but feels a lot easier to use.

Paredit with paredit-mode is a compelling plugin for editing elisp.